saving progress, change Event field price to double

+++ code/app/src/main/java/com/example/myapplication/model/Event.java
@@ -32,7 +32,7 @@ public class Event implements java.io.Serializable {
     private Date registrationEnd;
     private String location;
     private int capacity;
-    private int price;
+    private double price;
     private String posterUrl;
     private String qrCodeHash;
     private List<String> waitlist;
@@ -93,7 +93,7 @@ public class Event implements java.io.Serializable {
      * Constructors for creating a new Event
      */
     public Event(String eventName, String eventDescription, Date eventStart, Date eventEnd,
-                 Date registrationStart, Date registrationEnd, String location, int capacity, int price,
+                 Date registrationStart, Date registrationEnd, String location, int capacity, double price,
                  String posterUrl, String qrCodeHash, String organizerId) {
         this.eventName = eventName;
         this.eventDescription = eventDescription;
@@ -118,7 +118,7 @@ public class Event implements java.io.Serializable {
      * Constructors for creating a new Event, INCLUDING eventID
      */
     public Event(String eventId,String eventName, String eventDescription, Date eventStart, Date eventEnd,
-                 Date registrationStart, Date registrationEnd,String location, int capacity, int price,
+                 Date registrationStart, Date registrationEnd,String location, int capacity, double price,
                  String posterUrl, String qrCodeHash, String organizerId) {
         this.eventId = eventId;
         this.eventName = eventName;
@@ -350,7 +350,7 @@ public class Event implements java.io.Serializable {
      * Get the event price
      * @return the event price
      */
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
@@ -485,7 +485,7 @@ public class Event implements java.io.Serializable {
      * Set the event price
      * @param price
      */
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
         events.document(eventId).update("price", price);
     }
